requirements.txt
pyreadstat>=1.3
pandas>=2.0
pyyaml>=6.0
mlflow>=2.9
lightgbm>=4.0
numpy>=1.23
pyarrow>=14.0
scikit-learn>=1.3
psutil>=5.9
matplotlib>=3.7
seaborn>=0.12
pyarrow>=12.0
